enhancement: seperate function, print game info to console #5
